npu: concurrent multi-core matmul (ORK_NPU_MC=n) — 1.69x decode on RK3588
Splits each N-slice's output tiles across n cores (worker thread per core, per-core
regcmd/task/feature/output buffers), submits concurrently (core_mask=1<<i, all subcore_task[]
populated), accumulates disjoint columns of cres lock-free. n is a runtime request, clamped to
soc->cores (and ORK_MAXCORE=4) so it can never exceed the hardware — RK3576=2, RK3588=3, future
parts more, and the caller may pick fewer.

Measured (Qwen3-1.7B int8, RK3588, DDR 2112): decode 5.08 (1c) -> 6.46 (2c) -> 8.59 tok/s (3c,
1.69x), now ~75% of librkllmrt (11.5). Telemetry: all 3 cores active, RAM BW peak 42%->88%.
Correct: test_matmul ALL OK at ORK_NPU_MC=3. (rknpu_ioctl.h __u64/__s64 -> kernel's exact
underlying types so <pthread.h>'s transitive defs are an identical, legal redefinition.)

+++ src/rknpu_ioctl.h
@@ -15,13 +15,16 @@
 #include <stdint.h>
 #include <sys/ioctl.h>
 
-/* Self-contained kernel types (avoid a hard dep on linux/types.h). */
+/* Self-contained kernel types (avoid a hard dep on linux/types.h). Use the kernel's exact
+ * underlying types (asm-generic/int-ll64.h) so that if another header (e.g. pulled transitively
+ * by <pthread.h>) also defines __u64/__s64, it is an identical — and therefore legal — typedef
+ * redefinition, not a `unsigned long` vs `unsigned long long` conflict. Same sizes either way. */
 #ifndef __rknpu_kernel_types
 #define __rknpu_kernel_types
-typedef uint32_t __u32;
-typedef int32_t  __s32;
-typedef uint64_t __u64;
-typedef int64_t  __s64;
+typedef unsigned int       __u32;
+typedef int                __s32;
+typedef unsigned long long __u64;
+typedef long long          __s64;
 #endif
 
 /* ---- Minimal DRM ioctl encoding (from <drm/drm.h>) ---- */
